Rename api to endpoints to Avoid Confusion
