+++ server.go
@@ -4,7 +4,7 @@ import (
 	"embed"
 	"fmt"
 
-	"github.com/Leagueify/api/internal/api"
+	"github.com/Leagueify/api/internal/endpoints"
 	"github.com/Leagueify/api/internal/config"
 	"github.com/getsentry/sentry-go"
 	sentryecho "github.com/getsentry/sentry-go/echo"
